Main_computer_codes/requirements.txt
pyserial==3.5
requests==2.31.0
numpy==1.24.3
